requirements.txt
scipy
pandas
requests
numpy
python-dotenv
upstox-python-sdk
colorama
matplotlib
